requirements.txt
# Web Framework
fastapi[standard]==0.115.0
uvicorn[standard]==0.27.0
scalar-fastapi==1.0.3

# Database
sqlalchemy[asyncio]==2.0.36
# Using psycopg3 (has Python 3.13 wheels)
psycopg[binary]==3.2.3
# asyncpg==0.30.0  # Temporarily disabled - needs C++ compiler on Windows with Python 3.13
alembic==1.13.1
redis[hiredis]==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.0.1

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.1.0


joblib==1.3.2

# Email Service
resend==0.8.0

# Utilities
python-dotenv==1.0.0
ipython==8.20.0
requests==2.31.0
